Add approval-joins to group model
This adds an APPROVED role and implements it accordingly in model
methods as well as testing/code.
To have full access to a group, a user has to be approved. As the
current code just calls add_member, which adds the role directly,
nothing changes.
A next step would be to use add_applicant instead and add API/process
for approving a member to become an actual member.

This also refactors some direct usages of a GroupMembership to instead
work with convenience helpers.