Use yapf for consistent formatting (#524)
Ok, I reworked this now on-top of master, and included `./yapf.py` script to run it (with our excludes).

In the future if https://github.com/google/yapf/pull/543 gets merged we can use a yapf ignore file instead.

So, the plan would be:
- merge this to master
- cherry pick the "add yapf config" commit into any PR branch you might want to merge in the future
- run `./yapf.py` on your branch
- hopefully when you come to merge it'll all work nicely :)

Until we have gone through all the active branches like this, we should not manually reformat too much, or change the yapf config... keeps the pain down a bit. After that maybe some cleanup phases to (for example add lots of trailing commas).