Update setup.py
```
        exclude=['config/local_settings'],  # doesn't work :(
```

it is not working because config/local_settings is not a package but a *file* in a package; if it were a subpackage in config, then ``config.local_settings`` would work.
See: https://github.com/pypa/setuptools/blob/a4dbe3457d89cf67ee3aa571fdb149e6eb544e88/setuptools/__init__.py#L89
I would suggest to include a generic config.local-settings file when the package is built (isn't that the case anyway?) and the package should not be built on a local (dev) machine anyways, right? So I just deleted this line to write you this message … you don't have to merge it.